api/requirements.txt
supabase>=2.3.0
google-generativeai>=0.8.0
pyjwt>=2.8.0
